cloud-functions/email-processor/requirements.txt
functions-framework==3.5.0
google-cloud-storage==2.10.0
requests==2.31.0